Cleanup, make optional props

+++ ui/v2.5/src/components/Galleries/GalleryDetails/GalleryEditPanel.tsx
@@ -39,7 +39,7 @@ import { galleryTitle } from "src/core/galleries";
 
 interface IProps {
   isVisible: boolean;
-  onDelete: () => void;
+  onDelete?: () => void;
 }
 
 interface INewProps {
@@ -133,9 +133,11 @@ export const GalleryEditPanel: React.FC<
       Mousetrap.bind("s s", () => {
         formik.handleSubmit();
       });
-      Mousetrap.bind("d d", () => {
-        onDelete();
-      });
+      if (onDelete) {
+        Mousetrap.bind("d d", () => {
+          onDelete();
+        });
+      }
 
       // numeric keypresses get caught by jwplayer, so blur the element
       // if the rating sequence is started
@@ -163,7 +165,9 @@ export const GalleryEditPanel: React.FC<
 
       return () => {
         Mousetrap.unbind("s s");
-        Mousetrap.unbind("d d");
+        if (onDelete) {
+          Mousetrap.unbind("d d");
+        }
 
         Mousetrap.unbind("r");
       };
@@ -438,13 +442,15 @@ export const GalleryEditPanel: React.FC<
             >
               <FormattedMessage id="actions.save" />
             </Button>
-            <Button
-              className="edit-button"
-              variant="danger"
-              onClick={() => onDelete()}
-            >
-              <FormattedMessage id="actions.delete" />
-            </Button>
+            {onDelete && (
+              <Button
+                className="edit-button"
+                variant="danger"
+                onClick={() => onDelete()}
+              >
+                <FormattedMessage id="actions.delete" />
+              </Button>
+            )}
           </div>
           <Col xs={6} className="text-right">
             {renderScraperMenu()}
